+++ shell-zsh/.config/zsh/aliases.zsh
@@ -1,0 +1,8 @@
+alias arch='archaeology-0-core'
+alias arch0='archaeology-0-core'
+
+# Archaeology aliases
+archtime='archaeology-0-core --timeline'
+archwk='archaeology-0-core --this-week'
+archint='archaeology-0-core --by-intent'
+archsince='archaeology-0-core --since'